# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_bigops_platform_internal_model.User:
    properties:
      avatar:
        type: string
      created_at:
        example: "2024-01-01 00:00:00"
        type: string
      email:
        type: string
      id:
        type: integer
      phone:
        type: string
      real_name:
        type: string
      status:
        type: integer
      updated_at:
        example: "2024-01-01 00:00:00"
        type: string
      username:
        type: string
    type: object
...
  github_com_bigops_platform_internal_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  github_com_bigops_platform_internal_service.LoginResult:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/github_com_bigops_platform_internal_model.User'
    type: object
  internal_handler.ChangePasswordRequest:
    properties:
      new_password:
        example: NewPass456
        maxLength: 50
        minLength: 6
        type: string
      old_password:
        example: OldPass123
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  internal_handler.LoginRequest:
    properties:
      password:
        example: Admin123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  internal_handler.RegisterRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: Pass123456
        maxLength: 50
        minLength: 6
        type: string
      username:
        example: johndoe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: BigOps 运维平台 API 文档
  title: BigOps API
  version: "1.0"
paths:
...
  /auth/info:
    get:
      description: 根据 token 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_bigops_platform_internal_model.User'
              type: object
        "401":
          description: 用户未认证
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取当前登录用户信息
      tags:
      - 认证
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名密码登录，返回 JWT token 和用户信息
      parameters:
      - description: 登录请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_bigops_platform_internal_service.LoginResult'
              type: object
        "401":
          description: 用户名或密码错误/账号已被禁用
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
      summary: 用户登录
      tags:
      - 认证
  /auth/logout:
    post:
      description: 使当前 token 失效，加入 Redis 黑名单
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
        "400":
          description: token 格式错误
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - 认证
  /auth/password:
    post:
      consumes:
      - application/json
      description: 验证旧密码后修改为新密码，新密码需包含大小写字母和数字，长度 8-50 位
      parameters:
      - description: 修改密码请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 密码修改成功
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
        "400":
          description: 参数错误/原密码错误/密码复杂度不够
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 修改当前用户密码
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 注册新用户账号，密码需包含大小写字母和数字，长度 8-50 位
      parameters:
      - description: 注册请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
        "400":
          description: 参数错误/用户名已存在
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
      summary: 用户注册
      tags:
      - 认证
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token 认证，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"